♻️ MyGroup 컴포넌트 디렉토리로 분리하고 스타일 파일도 분리해서 사용
