+++ app/frontend/src/pages/Profile/MyGroup/index.tsx
