+++ app/frontend/src/pages/Profile/index.css.ts
@@ -1,7 +1,7 @@
 import { style } from '@vanilla-extract/css';
 
 import { vars } from '@/styles';
-import { sansBold16, sansRegular18 } from '@/styles/font.css';
+import { sansRegular18 } from '@/styles/font.css';
 
 export const container = style({
   display: 'flex',
@@ -11,21 +11,6 @@ export const container = style({
   margin: '0 auto 8rem auto',
 });
 
-export const groupButtons = style({
-  display: 'flex',
-  gap: '1.2rem',
-});
-export const groupListButton = style([
-  sansBold16,
-  {
-    display: 'flex',
-    gap: 0,
-    alignItems: 'center',
-    alignSelf: 'end',
-    color: vars.color.morakGreen,
-  },
-]);
-
 export const list = style({
   display: 'flex',
   flexDirection: 'column',
@@ -34,22 +19,8 @@ export const list = style({
   overflowY: 'auto',
 });
 
-export const loading = style({
-  width: '100%',
-  display: 'flex',
-  justifyContent: 'center',
-});
-
 export const logoutButton = style({
   alignSelf: 'end',
-});
-
-export const navLink = style({
-  width: '100%',
-});
-
-export const rotateArrow = style({
-  transform: 'rotate(180deg)',
 });
 
 export const section = style({
